+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.4.11 - 12 Oct 2025
+Bug fix:
+- [#1469](https://github.com/elysiajs/elysia/issues/1469) incorrect ping, pong type signature
+- make `file-type` non optional dependency to fix default build problem
+
 # 1.4.10 - 9 Oct 2025
 Bug fix:
 - [#1406](https://github.com/elysiajs/elysia/issues/1406) enforce return type in OptionalHandler